requirements.txt
2captcha-python==1.1.2
alembic==1.7.5
anyio==3.5.0
api==0.0.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astroid==2.9.2
asttokens==2.0.5
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.1
entrypoints==0.4
executing==0.8.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nose==1.3.7
notebook==6.4.8
outcome==1.1.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.11.2
pylint==2.12.2
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
pyshorteners==1.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.4.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==225
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.27.1
selenium==4.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.30
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
